validation sub-functions now receive and return value.
sublclassing was too difficult when the user had to know where to find
the value for different stages. Now, the value to be validated is
passed into the function, and the value to save is returned by the
function.

you can now get a list of all errors with getErrors().