Fixed lots of bugs; simplified skinning
got rid of skinned fields. now skin is a string. when a skin is
specified, it is inherited by all children. when a field creates
its widget, searches for a widget named widget.<skin>.<widgetname>.
if it doesn't find one, it just uses widget.<widgetname>. regardless
of the widget kind, the widget callse <skin>Skin() to actually
generate a custom look for the widget. if the custom skin function is
not found it defaults to calling Skin().

containerFields no longer explicitly set the value. now that
we have `initializeSubfield` it is not necessary, as values are
set on widget at initilization. However, ListFields now call
setValue whenever setSchema is called.

Almost all subfield functionality has been moved from the widget
to the field. the field assumes that subfields will be stored in
this.$.widget.$.fields, if a widget exists. `resetFields` now
destroys widget's components, and `setValue` creates the widget
components and renders. ListWidget accomplishes this by calling
addField on either the widget or the field, if widget doesn't exist.

resetFields also stores the value of existing fields so widgets
can be changed at runtime.

`widgetChanged` no longer modifies the widget attribute, instead
creating a copy and modifying that. this makes it possible in future
to modify widget at runtime. for this reason, prepareWidget now
accepts this local modified widget and must return said widget.