abandoned bootstrap skin; created onyx skin; bug fixes
Realized that the bootstrap positional css was becoming impossibly
unwealdy, so abandoned it. onyx skin based on bootstrap layout
created quickly.

Added validation strategy, instant validation and widgets selectors
to sample app.

now, when an inheritable property is set on a field, it causes
a recreation of widget and all subfield widgets.