Fixed widget add/delete and dynamic schema
everything is now done by the field, and the widgets just listen.
they listen for field creation and create a field, they listen for
deletion and do deletion. they listen for value change...
Nothing that impacts the fields is done by a widget except changed
values of the widget. Instead, the widget sends a message to the
field that something needs to be done (like add/remove field)
the widget passes a message