simplified containerField init; getValue always returns value
Field init now calls setSchema if it exists, in the proper place,
rather than trying to do field init, then containerField init, then
redo field init that should have happened after setSchema.

before any change in value to a containerfield, the @value is set to
the current value. then, getValue always checks to see if @value is
defined and if so returns that. That way, when responding to onFieldAdd
events, the value of the parent container will always be valid, rather
than the value of whatever subfields happen to have been instantiated
to that point.